Implemented tests for users route and fixed bugs

+++ routes/users.js
@@ -12,11 +12,16 @@ const bodyParser = require('koa-bodyparser');
 const userModel = require('../models/users');
 const favsModel = require('../models/favourites');
 const chatModel = require('../models/chats');
+const dogModel = require('../models/dogs');
 
 const { auth } = require('../controllers/auth');
 const { config } = require('../config');
 
-const { validateUser, validateFavourite } = require('../controllers/validation');
+const {
+	validateUser,
+	validateUserUpdate,
+	validateFavourite
+} = require('../controllers/validation');
 
 const can = require('../permissions/users');
 const { clamp } = require('../helpers/utils');
@@ -29,7 +34,7 @@ router.use(bodyParser());
  * @param {object} ctx context passed from Koa.
  */
 const getAll = async ctx => {
-	const permission = await can.getAll(ctx.state.user);
+	const permission = await can.user.getAll(ctx.state.user);
 	if (!permission.granted) {
 		ctx.status = 403;
 		return;
@@ -50,7 +55,7 @@ const getAll = async ctx => {
 		users = users.map(user => permission.filter(user));
 		ctx.body = users;
 	} catch (err) {
-		ctx.status = 500;
+		ctx.status = 400;
 		ctx.body = err;
 	}
 };
@@ -60,14 +65,14 @@ const getAll = async ctx => {
  * @param {object} ctx context passed from Koa.
  */
 const getUser = async ctx => {
-	const id = ctx.params.id;
+	const id = parseInt(ctx.params.id);
 	let { select = [] } = ctx.request.query;
 	if (!Array.isArray(select)) select = Array(select);
 	try {
 		let user = await userModel.getById(id, select);
 		if (user) {
 			const { id: userId, role } = ctx.state.user;
-			const permission = await can.get(role, userId, id);
+			const permission = await can.user.get(role, userId, id);
 			if (!permission.granted) {
 				ctx.status = 403;
 				return;
@@ -76,7 +81,7 @@ const getUser = async ctx => {
 			ctx.body = user;
 		}
 	} catch (err) {
-		ctx.status = 500;
+		ctx.status = 400;
 		ctx.body = err;
 	}
 };
@@ -86,19 +91,27 @@ const getUser = async ctx => {
  * @param {object} ctx context passed from Koa.
  */
 const createUser = async ctx => {
-	const { staffKey, role, ...body } = ctx.request.body;
+	const { staffKey, ...body } = ctx.request.body;
+	const { username, email } = body;
 	// Giving the user the role 'staff' if they provide right key.
 	body.role = staffKey === config.staffKey ? 'staff' : 'user';
-	try {
-		const id = await userModel.add(body);
-		if (id) {
-			ctx.status = 201;
-			ctx.body = { ID: id, created: true, link: `${ctx.request.path}/${id}` };
-		}
-	} catch (err) {
-		ctx.status = 500;
-		ctx.body = err;
+	// Making sure username isn't taken
+	let user = await userModel.getByUsername(username);
+	if (user) {
+		ctx.status = 400;
+		ctx.body = { message: 'Username is taken.' };
+		return;
 	}
+	// Making sure email isn't taken
+	user = await userModel.getByEmail(email);
+	if (user) {
+		ctx.status = 400;
+		ctx.body = { message: 'Email is taken.' };
+		return;
+	}
+	const id = await userModel.add(body);
+	ctx.status = 201;
+	ctx.body = { id, created: true, link: `${ctx.request.path}/${id}` };
 };
 
 /**
@@ -106,19 +119,18 @@ const createUser = async ctx => {
  * @param {object} ctx context passed from Koa.
  */
 const updateUser = async ctx => {
-	const userId = ctx.params.id;
-	const user = await userModel.getById(userId);
+	const { id: userId, role } = ctx.state.user;
+	const id = parseInt(ctx.params.id);
+	const user = await userModel.getById(id, []);
 	if (user) {
-		// Excluding fields that must not be updated
-		const { id, dateCreated, ...body } = ctx.request.body;
-		try {
-			const result = await userModel.update(userId, body);
-			// Knex returns amount of affected rows.
-			if (result) ctx.body = { id: userId, updated: true, link: ctx.request.path };
-		} catch (err) {
-			ctx.status = 500;
-			ctx.body = err;
+		const permission = await can.user.modify(role, userId, id);
+		if (!permission.granted) {
+			ctx.status = 403;
+			return;
 		}
+		const { body } = ctx.request;
+		await userModel.update(id, body);
+		ctx.body = { id: id, updated: true, link: ctx.request.path };
 	}
 };
 
@@ -127,16 +139,17 @@ const updateUser = async ctx => {
  * @param {object} ctx context passed from Koa.
  */
 const deleteUser = async ctx => {
-	const id = ctx.params.id;
-	try {
-		const user = await userModel.getById(id);
-		if (user) {
-			const result = await userModel.delete(id);
-			if (result) ctx.body = { id, deleted: true };
+	const id = parseInt(ctx.params.id);
+	const { id: userId, role } = ctx.state.user;
+	const user = await userModel.getById(id, []);
+	if (user) {
+		const permission = await can.user.delete(role, userId, id);
+		if (!permission.granted) {
+			ctx.status = 403;
+			return;
 		}
-	} catch (err) {
-		ctx.status = 500;
-		ctx.body = err;
+		await userModel.delete(id);
+		ctx.body = { id, deleted: true };
 	}
 };
 
@@ -145,13 +158,17 @@ const deleteUser = async ctx => {
  * @param {object} ctx context passed from Koa.
  */
 const getUserFavs = async ctx => {
-	const id = ctx.params.id;
-	try {
+	const id = parseInt(ctx.params.id);
+	const { id: userId, role } = ctx.state.user;
+	const user = await userModel.getById(id, []);
+	if (user) {
+		const permission = await can.favourite.get(role, userId, id);
+		if (!permission.granted) {
+			ctx.status = 403;
+			return;
+		}
 		const favourites = await favsModel.getByUserId(id);
-		if (favourites) ctx.body = favourites;
-	} catch (err) {
-		ctx.status = 500;
-		ctx.body = err;
+		ctx.body = favourites;
 	}
 };
 
@@ -160,21 +177,36 @@ const getUserFavs = async ctx => {
  * @param {object} ctx context passed from Koa.
  */
 const addUserFav = async ctx => {
-	const userId = ctx.params.id;
+	const id = parseInt(ctx.params.id);
+	const { id: userId, role } = ctx.state.user;
 	const { dogId } = ctx.request.body;
-	try {
-		const result = await favsModel.add(userId, dogId);
-		if (result) {
+	const user = await userModel.getById(id, []);
+	if (user) {
+		const permission = await can.favourite.create(role, userId, id);
+		if (!permission.granted) {
+			ctx.status = 403;
+			return;
+		}
+		const dog = await dogModel.getById(dogId, []);
+		if (dog) {
+			let favs = await favsModel.getByUserId(id);
+			favs = favs.filter(fav => fav.dogId === dogId);
+			if (favs.length) {
+				ctx.status = 400;
+				ctx.body = { message: 'Dog is already a favourite.' };
+				return;
+			}
+			const favId = await favsModel.add(userId, dogId);
 			ctx.status = 201;
 			ctx.body = {
-				id: userId,
+				id: favId,
 				created: true,
-				link: `${ctx.request.path}/${dogId}`
+				link: `${ctx.request.path}/${favId}`
 			};
+		} else {
+			ctx.status = 400;
+			ctx.body = { message: 'Dog does not exist.' };
 		}
-	} catch (err) {
-		ctx.status = 500;
-		ctx.body = err;
 	}
 };
 
@@ -183,13 +215,21 @@ const addUserFav = async ctx => {
  * @param {object} ctx context passed from Koa.
  */
 const getUserFav = async ctx => {
-	const favId = ctx.params.favId;
-	try {
-		const result = await favsModel.getSingleFav(favId);
-		if (result) ctx.body = result;
-	} catch (err) {
-		ctx.status = 500;
-		ctx.body = err;
+	let { id, favId } = ctx.params;
+	id = parseInt(id);
+	favId = parseInt(favId);
+	const { id: userId, role } = ctx.state.user;
+	const user = await userModel.getById(id, []);
+	if (user) {
+		const favourite = await favsModel.getSingleFav(favId);
+		if (favourite) {
+			const permission = await can.favourite.get(role, userId, id);
+			if (!permission.granted) {
+				ctx.status = 403;
+				return;
+			}
+			ctx.body = favourite;
+		}
 	}
 };
 
@@ -198,13 +238,22 @@ const getUserFav = async ctx => {
  * @param {object} ctx context passed from Koa.
  */
 const deleteUserFav = async ctx => {
-	const favId = ctx.params.favId;
-	try {
-		const result = await favsModel.delete(favId);
-		if (result) ctx.body = { id: favId, deleted: true };
-	} catch (err) {
-		ctx.status = 500;
-		ctx.body = err;
+	let { id, favId } = ctx.params;
+	id = parseInt(id);
+	favId = parseInt(favId);
+	const { id: userId, role } = ctx.state.user;
+	const user = await userModel.getById(id, []);
+	if (user) {
+		const favourite = await favsModel.getSingleFav(favId);
+		if (favourite) {
+			const permission = await can.favourite.delete(role, userId, id);
+			if (!permission.granted) {
+				ctx.status = 403;
+				return;
+			}
+			await favsModel.delete(favId);
+			ctx.body = { id: favId, deleted: true };
+		}
 	}
 };
 
@@ -213,13 +262,17 @@ const deleteUserFav = async ctx => {
  * @param {object} ctx context passed from koa.
  */
 const getUserChats = async ctx => {
-	const userId = ctx.params.id;
-	try {
+	const userId = parseInt(ctx.params.id);
+	const { id, role } = ctx.state.user;
+	const user = await userModel.getById(userId, []);
+	if (user) {
+		const permission = await can.user.get(role, userId, id);
+		if (!permission.granted) {
+			ctx.status = 403;
+			return;
+		}
 		const chats = await chatModel.getByUserId(userId);
-		if (chats) ctx.body = chats;
-	} catch (err) {
-		ctx.status = 500;
-		ctx.body = err;
+		ctx.body = chats;
 	}
 };
 
@@ -227,7 +280,7 @@ router.get('/', auth, getAll);
 router.post('/', validateUser, createUser);
 
 router.get('/:id([0-9]+)', auth, getUser);
-router.put('/:id([0-9]+)', auth, validateUser, updateUser);
+router.put('/:id([0-9]+)', auth, validateUserUpdate, updateUser);
 router.del('/:id([0-9]+)', auth, deleteUser);
 
 router.get('/:id([0-9]+)/favourites', auth, getUserFavs);
